requirements.txt
beautifulsoup4
lxml
requests
aiohttp
python-dotenv